\name{Impustep-package}
\alias{Impustep-package}
\alias{Impustep}
\docType{package}
\title{
\packageTitle{Impustep}
}
\description{
\packageDescription{Impustep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Impustep}
\packageIndices{Impustep}
Functions belonging to the "5 steps for imputation" paper.
}
\author{
\packageAuthor{Benjamin Gravesteijn}

Maintainer: \packageMaintainer{Benjamin Gravesteijn: b.gravesteijn@erasmusmc.nl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
